Split PDG terminal guidance into partial class
