+++ MechJeb2/MechJeb2.csproj
@@ -75,6 +75,7 @@
     <Compile Include="PoweredDescentGuidance\PDGTypes.cs" />
     <Compile Include="PoweredDescentGuidance\PDGMathUtils.cs" />
     <Compile Include="PoweredDescentGuidance\PDGEngine.cs" />
+    <Compile Include="PoweredDescentGuidance\PDGTerminal.cs" />
     <Compile Include="Maneuver\Operation.cs" />
     <Compile Include="Maneuver\OperationAdvancedTransfer.cs" />
     <Compile Include="Maneuver\OperationApoapsis.cs" />
